+++ pages/[category].tsx
@@ -1,5 +1,3 @@
-import Nav from '../components/Nav';
-import Footer from '../components/Footer';
 import Category from '../containers/Category';
 import { useRouter } from 'next/router';
