+++ pages/_app.tsx
@@ -1,42 +1,11 @@
-import { Global, css } from '@emotion/react';
+import { Global } from '@emotion/react';
 import { AppProps } from 'next/app';
 import { RecoilRoot } from 'recoil';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { queryClient } from '../utils/queryClient';
 import Nav from '../components/Nav';
 import Footer from '../components/Footer';
-
-const globalStyles = css`
-  * {
-    box-sizing: border-box;
-  }
-
-  ul,
-  li {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-  }
-
-  label:hover {
-    font-weight: 900;
-    color: #0b57d0;
-  }
-
-  input:checked + span {
-    font-weight: 900;
-    color: #0b57d0;
-  }
-
-  button {
-    border: 1px solid rgb(230, 230, 230);
-    background-color: rgb(230, 230, 230);
-  }
-
-  button:hover {
-    border: 1px solid black;
-  }
-`;
+import globalStyles from '../styles/globalStyles';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
