+++ styles/filterStyles.ts
@@ -1,0 +1,11 @@
+import { css } from '@emotion/react';
+
+export const filterStyles = css`
+  display: inline;
+  line-height: 50px;
+  font-size: 10px;
+  padding: 0 5px;
+  & > input {
+    display: none;
+  }
+`;
