+++ styles/mainStyles.ts
@@ -1,0 +1,16 @@
+import { css } from '@emotion/react';
+
+export const mainStyle = css`
+  padding: 50px;
+`;
+
+export const filtersStyle = css`
+  height: 50px;
+  font-size: 10px;
+`;
+
+export const productsStyle = css`
+  display: grid;
+  grid-template-columns: repeat(auto-fill, 250px);
+  font-size: 10px;
+`;
